fixed bus deletion bug that causes assigned trip to be removed

+++ BTRS/Controllers/BusController.cs
@@ -51,12 +51,26 @@ namespace BTRS.Controllers
 
         public ActionResult Delete(int id)
         {
-            Bus bus = _context.bus.Find(id);
-            _context.bus.Remove(bus);
-            _context.SaveChanges();
+            Bus bus = _context.bus.Include(b => b.trips).FirstOrDefault(b => b.BusId == id);
+
+            if (bus == null)
+            {
+                return NotFound();
+            }
+
+            if (bus.trips != null && bus.trips.Count > 0)
+            {
+                TempData["msg"] = "Cannot remove a bus when it is assigned to a trip";
+            }
+            else
+            {
+                _context.bus.Remove(bus);
+                _context.SaveChanges();
+            }
 
             return RedirectToAction(nameof(Index));
         }
+
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -64,7 +78,6 @@ namespace BTRS.Controllers
         {
             try
             {
-
                 return RedirectToAction(nameof(Index));
             }
             catch
